requirements.txt
fastapi[all] == 0.110.0
starlette ==0.36.3
bcrypt == 4.1.2
passlib[bcrypt] == 1.7.4
python-jose == 3.3.0 
numpy==1.26.4
pandas == 2.2.2
SQLAlchemy==2.0.30
pytest==7.4.4
httpx==0.27.0
pytest-asyncio==0.23.5